+++ lib/fluent_assertions.dart
@@ -1,7 +1,8 @@
-/// Support for doing something awesome.
+/// The Fluent Assertions library written in Dart.
 ///
-/// More dartdocs go here.
+/// It uses Dart's Extension Functions to provide a fluent wrapper around test assertions.
 library fluent_assertions;
 
 export 'src/basic_assertions.dart';
 export 'src/numerical_assertions.dart';
+export 'src/string_assertions.dart';